Create workingRTC
The RTC is working in this code. The channel part in the void loop is not completely updated. The RTC turns on five minutes before the client is supposed to change the channel. The LED remains on until the client presses the button to change the channel. Once the button is pressed the LED turns off. If the button is not pressed the LED will remain on. When the remote is connected the clock will need to be set for the specific time.