Added Sleep Mode (Untested)
Added sample sleep mode code from: https://playground.arduino.cc/Learning/ArduinoSleepCode
Did not change pins, which need to be modified. I do not remember which pin we are using for interrupts, if it is pin 2, we are fine. If its pin 3, we need to modify the functions that the sleep library gives (namely attachInterrupt).